made switch

+++ node2/drivers/include/UARTDriverIoT.h
@@ -1,2 +1,2 @@
-void USART_Transmit_STXETX( uint32_t payload, uint8_t firstByte );
+void USART_Transmit_STXETX( uint8_t payload, uint8_t firstByte );
 uint8_t* USART_Receive_STXETX();
